--- lib/game_board/zacatrus/domain/details_page/game_review.dart
@@ -1,15 +1,0 @@
-class GameReview {
-  GameReview({
-    this.title,
-    this.stars,
-    this.description,
-    this.author,
-    this.date,
-  });
-
-  String? title;
-  double? stars;
-  String? description;
-  String? author;
-  String? date;
-}
